Fix growing window, remove name popover
This commit removes name popovers which shows when the user hovers over
guild icons. This seems to fix a bug where the window randomly freezes
on GNOME.

This also possibly fixes a bug where an image attachment will cause the
window to grow horizontally across the whole screen. This is relevant to
issue #46.