Prohibit large file uploads
Previously, Dissent would let you choose files as large as you want.
This commit adds a new alert dialog that warns the user about the file
size limit and prevents them from being added if they exceeed that
limit.

We suspect that this was causing account warnings, so hopefully this
lets users be more mindful.

Fixes #350.